rename PowerFactoryGrid to RawGridModel
