+++ src/main/scala/edu/ie3/powerFactory2psdm/util/GridPreparator.scala
@@ -7,30 +7,31 @@
 package edu.ie3.powerFactory2psdm.util
 
 import com.typesafe.scalalogging.LazyLogging
-import edu.ie3.powerFactory2psdm.model.powerfactory.PowerFactoryGrid
-import edu.ie3.powerFactory2psdm.model.powerfactory.PowerFactoryGrid.Switches
+import edu.ie3.powerFactory2psdm.model.powerfactory.RawGridModel
+import edu.ie3.powerFactory2psdm.model.powerfactory.RawGridModel.Switches
 
 object GridPreparator extends LazyLogging {
 
   /**
-    * Perform preparation of the [[PowerFactoryGrid]] before the actual conversion can happen.
+    * Perform preparation of the [[RawGridModel]] before the actual conversion can happen.
     *
-    * @param pfGrid the [[PowerFactoryGrid]] to prepare
-    * @return the prepared [[PowerFactoryGrid]]
+    * @param pfGrid the [[RawGridModel]] to prepare
+    * @return the prepared [[RawGridModel]]
     */
-  def prepare(pfGrid: PowerFactoryGrid): PowerFactoryGrid = {
+  def prepare(pfGrid: RawGridModel): RawGridModel = {
     val filteredSwitches = removeSinglyConnectedSwitches(pfGrid.switches)
     pfGrid.copy(switches = filteredSwitches)
   }
 
   /**
-    * Removes [[Switches]] from a [[PowerFactoryGrid]] that are only connected to a single node.
-    * @param maybeSwitches
+    * Removes [[Switches]] from a [[RawGridModel]] that are only connected to a single node.
+   *
+   * @param maybeSwitches
     * @return
     */
   def removeSinglyConnectedSwitches(
-      maybeSwitches: Option[List[PowerFactoryGrid.Switches]]
-  ): Option[List[PowerFactoryGrid.Switches]] =
+      maybeSwitches: Option[List[RawGridModel.Switches]]
+  ): Option[List[RawGridModel.Switches]] =
     maybeSwitches.map(switches => {
       val (fullyConnected, singlyConnected) =
         switches.partition(isFullyConnectedSwitch)
